Remove superfluous status metadata
This repository was copied and adapted from the GDS/GOV.UK one, where they handle lots of RFCs and some complexity we don't currently have.

The status metadata is confusing in its current state, where things end up in the main branch with 'proposed' even though they're accepted, and it's not much better to set it as accepted in the branch (as it isn't until it's been merged).

Rather than overlaying multiple ways to track status, this PR removed the status metadata so we only use one criteria:

> If it's in the main branch, it's accepted.
> If it's in a PR, it's proposed
> Anything else is draft